// front-vue/src/types.ts
export type TranslatorKey =
  | 'youdao' | 'baidu' | 'deepl' | 'papago' | 'caiyun' | 'sakura' | 'offline'
  | 'openai' | 'deepseek' | 'groq' | 'gemini' | 'custom_openai'
  | 'nllb' | 'nllb_big' | 'sugoi' | 'jparacrawl' | 'jparacrawl_big'
  | 'm2m100' | 'm2m100_big' | 'mbart50' | 'qwen2' | 'qwen2_big' | 'none';